# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AdminGuidancePostDto" : {
        "properties" : {
          "bodyHtml" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdBy" : {
            "format" : "int64",
            "type" : "integer"
          },
          "heroImageAlt" : {
            "type" : "string"
          },
          "heroImageId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "heroImageUrl" : {
            "type" : "string"
          },
          "heroImportUrl" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "locale" : {
            "type" : "string"
          },
          "pinned" : {
            "type" : "boolean"
          },
          "slug" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "DRAFT", "PUBLISHED" ],
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "properties" : {
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "GuidancePostDto" : {
        "properties" : {
          "alternates" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "type" : "object"
          },
          "bodyHtml" : {
            "type" : "string"
          },
          "heroImageAlt" : {
            "type" : "string"
          },
          "heroImageUrl" : {
            "type" : "string"
          },
          "locale" : {
            "type" : "string"
          },
          "localeFallback" : {
            "type" : "boolean"
          },
          "pinned" : {
            "type" : "boolean"
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UpdateGuidancePostRequest" : {
        "properties" : {
          "body" : {
            "type" : "string"
          },
          "heroImageAlt" : {
            "maxLength" : 300,
            "minLength" : 0,
            "type" : "string"
          },
          "heroImageId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "heroImportUrl" : {
            "maxLength" : 2048,
            "minLength" : 0,
            "type" : "string"
          },
          "locale" : {
            "maxLength" : 5,
            "minLength" : 0,
            "type" : "string"
          },
          "pinned" : {
            "type" : "boolean"
          },
          "slug" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "title" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "body", "title" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "The `Authorization: Bearer <accessToken>` header read by the JWT filter. A valid token of a SUSPENDED account authenticates nothing (fresh per-request DB lookup — the DB is the truth, never a token claim).",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "aleks.bratsun@reaktiiv.com",
      "name" : "Aleks Bratsun (project owner)"
    },
    "description" : "The machine-readable contract of the OpenShelter backend. Companion to the human contracts: context-and-tasks/05-shelter-api.puml (the read contract), the README 'API' section and docs/whitepaper.md. Served at /v3/api-docs in dev/test only (SPRINGDOC_ENABLED); the committed snapshot is docs/api/openapi.json. Every non-2xx response carries the uniform ErrorResponse body.",
    "license" : {
      "name" : "MIT"
    },
    "title" : "OpenShelter API",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/admin/guidance/{id}" : {
      "delete" : {
        "description" : "Requires confirm=true (400 without it). The post's media assets stay in the library (uploads are inventory, not garbage) and its audit rows keep their label snapshot. 204; 404 unknown id.",
        "operationId" : "delete_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Required: confirm=true.",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "confirm",
          "in" : "query",
          "description" : "Required: confirm=true.",
          "required" : false,
          "schema" : {
            "type" : "boolean",
            "default" : false
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Deleted"
          },
          "400" : {
            "description" : "confirm=true missing"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "description" : "Authenticated non-admin"
          },
          "404" : {
            "description" : "Unknown post id"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "summary" : "Hard delete of a guidance post",
        "tags" : [ "Admin guidance" ],
        "x-admin-only" : {
          "value" : "true"
        }
      },
      "get" : {
        "description" : "Id-keyed (a draft has a slug, but the admin form edits by id). 404 unknown id.",
        "operationId" : "get_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminGuidancePostDto"
                }
              }
            },
            "description" : "The post"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Validation failure — the uniform error body."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminGuidancePostDto"
                }
              }
            },
            "description" : "Authenticated non-admin"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminGuidancePostDto"
                }
              }
            },
            "description" : "Unknown post id"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "summary" : "The admin guidance detail",
        "tags" : [ "Admin guidance" ],
        "x-admin-only" : {
          "value" : "true"
        }
      },
      "put" : {
        "description" : "Full replace of the editable fields; the slug is kept when omitted (a given slug that another post holds → 409 naming it). The body is re-sanitized — the stored value is the sanitizer output. 200 with the updated post; 404 unknown id.",
        "operationId" : "update_2",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateGuidancePostRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminGuidancePostDto"
                }
              }
            },
            "description" : "The updated post"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminGuidancePostDto"
                }
              }
            },
            "description" : "Validation failure (incl. a pending import URL on a published post)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminGuidancePostDto"
                }
              }
            },
            "description" : "Authenticated non-admin"
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminGuidancePostDto"
                }
              }
            },
            "description" : "Unknown post id (or heroImageId with no such asset)"
          },
          "409" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminGuidancePostDto"
                }
              }
            },
            "description" : "Slug collision (naming the slug)"
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "summary" : "Full replace of a guidance post",
        "tags" : [ "Admin guidance" ],
        "x-admin-only" : {
          "value" : "true"
        }
      }
    },
...
    "/api/guidance" : {
      "get" : {
        "description" : "PUBLISHED only (drafts are invisible), ONE locale, pinned first, then publishedAt descending (id descending tie-break). The index does not carry the post body (bodyHtml is null). 200 with [] when nothing is published in the requested locale.",
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "locale",
          "in" : "query",
          "description" : "The reader's language (optional): only the published posts of this locale are returned. Absent = the server's default locale. Blank or more than 5 characters is a 400.",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/GuidancePostDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "The published posts of the requested locale (pinned first)"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/GuidancePostDto"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "A blank or over-long locale (the column is VARCHAR(5))"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Authenticated but not allowed (admin kind, verified account or authorship)."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unknown resource id — the uniform error body."
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "security" : [ ],
        "summary" : "The public guidance index",
        "tags" : [ "Public guidance" ]
      }
    },
    "/api/guidance/{slug}" : {
      "get" : {
        "description" : "PUBLISHED only, by slug. Serves the requested locale's translation, or the default-locale translation with localeFallback=true when the post has none in the requested locale (a 200, never a 404). Carries the stored (sanitized) bodyHtml and an alternates map (locale -> slug) for the language switcher.",
        "operationId" : "get_1",
        "parameters" : [ {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "locale",
          "in" : "query",
          "description" : "The reader's language (optional): 404 when the post is in another locale. Absent = the server's default locale. Blank or more than 5 characters is a 400.",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GuidancePostDto"
                }
              }
            },
            "description" : "The published post — the requested locale's translation, or the default-locale one (localeFallback=true) when the post lacks the requested locale"
          },
          "400" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GuidancePostDto"
                }
              }
            },
            "description" : "A blank or over-long locale (the column is VARCHAR(5))"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Missing/invalid/expired bearer token, or a suspended account whose tokens authenticate nothing."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Authenticated but not allowed (admin kind, verified account or authorship)."
          },
          "404" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GuidancePostDto"
                }
              }
            },
            "description" : "Unknown slug, a draft slug, or a post with neither a requested-locale nor a default-locale translation"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Conflict (duplicate or forbidden state) — the uniform error body."
          },
          "429" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Rate limit / report throttle exceeded — Retry-After in seconds."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Unexpected server error — the uniform error body."
          }
        },
        "security" : [ ],
        "summary" : "The public guidance detail",
        "tags" : [ "Public guidance" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
}